put fixtures in the right place
